+++ sklearn/utils/tests/test_validation.py
@@ -329,21 +329,24 @@ def test_check_array_accept_sparse_type_exception():
     X_csr = sp.csr_matrix(X)
     invalid_type = SVR()
 
-    msg = "A sparse matrix was passed, but dense data is required. " \
-        "Use X.toarray() to convert to a dense numpy array."
+    msg = ("A sparse matrix was passed, but dense data is required. "
+           "Use X.toarray() to convert to a dense numpy array.")
     assert_raise_message(TypeError, msg,
                          check_array, X_csr, accept_sparse=False)
     assert_raise_message(TypeError, msg,
                          check_array, X_csr, accept_sparse=None)
 
-    msg = "Parameter 'accept_sparse' should be a string, " \
-          "boolean or list of strings. You provided 'accept_sparse={}'."
+    msg = ("Parameter 'accept_sparse' should be a string, " 
+           "boolean or list of strings. You provided 'accept_sparse={}'.")
+    assert_raise_message(ValueError, msg.format(invalid_type),
+                         check_array, X_csr, accept_sparse=invalid_type)
+
+    msg = ("When providing 'accept_sparse' as a tuple or list, "
+           "it must contain at least one string value.")
     assert_raise_message(ValueError, msg.format([]),
                          check_array, X_csr, accept_sparse=[])
     assert_raise_message(ValueError, msg.format(()),
                          check_array, X_csr, accept_sparse=())
-    assert_raise_message(ValueError, msg.format(invalid_type),
-                         check_array, X_csr, accept_sparse=invalid_type)
 
     msg = "'SVR' object"
     assert_raise_message(TypeError, msg,